Add BassEngine abstraction
