+++ OsuPlayer/Core.cs
@@ -1,3 +1,4 @@
+using OsuPlayer.Modules.Audio.Engine;
 using OsuPlayer.Views;
 using OsuPlayer.Windows;
