+++ OsuPlayer/Modules/Audio/Engine/IAudioEngine.cs
@@ -1,0 +1,8 @@
+﻿using System.ComponentModel;
+
+namespace OsuPlayer.Modules.Audio.Engine;
+
+public interface IAudioEngine : ICanOpenFiles, IAudioControls, IHasEffects
+{
+    public event PropertyChangedEventHandler? PropertyChanged;
+}
